Split push into multiple modules
